Replace Otel Events with duration attributes
We want to be able to aggregate and calculate the duration for a
build/job object to transition from state A to state B.  We could not do
that with Otel events which are attached to a span as timestamp.

Instead, pre-calculate different state durations and attach them to the
span as custom attributes.  This should enable us to query for these
durations given a span.